address API done

+++ src/pages/AddressesPage.jsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from "react";
+import { useNavigate } from "react-router-dom";
 
 import MainLayout from "../layouts/MainLayout";
 
 import AddressForm from "../components/address/AddressForm";
-
 import AddressCard from "../components/address/AddressCard";
 
 import {
@@ -13,35 +13,61 @@ import {
   setDefaultAddress,
 } from "../api/addressApi";
 
-export default function AddressesPage() {
+export default function AddressPage() {
+  const navigate = useNavigate();
+
   const [addresses, setAddresses] = useState([]);
-
-  const loadAddresses = async () => {
-    const response = await getAddresses();
-
-    setAddresses(response.data.data);
-  };
 
   useEffect(() => {
     loadAddresses();
   }, []);
 
-  const handleCreate = async (form) => {
-    await createAddress(form);
+  const loadAddresses = async () => {
+    try {
+      const response = await getAddresses();
 
-    loadAddresses();
+      setAddresses(response.data.data);
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
+  const handleCreate = async (form) => {
+    try {
+      await createAddress(form);
+
+      await loadAddresses();
+    } catch (error) {
+      console.error(error);
+    }
   };
 
   const handleDelete = async (id) => {
-    await deleteAddress(id);
+    if (!window.confirm("Delete this address?")) {
+      return;
+    }
 
-    loadAddresses();
+    try {
+      await deleteAddress(id);
+
+      await loadAddresses();
+    } catch (error) {
+      console.error(error);
+    }
   };
 
   const handleDefault = async (id) => {
-    await setDefaultAddress(id);
+    try {
+      await setDefaultAddress(id);
 
-    loadAddresses();
+      await loadAddresses();
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
+  const handleEdit = (address) => {
+    navigate(`/addresses/${address.id}/edit`);
   };
 
   return (
@@ -50,15 +76,19 @@ export default function AddressesPage() {
         <div className="mx-auto max-w-6xl px-6">
           <h1 className="mb-10 text-4xl font-bold">My Addresses</h1>
 
-          <AddressForm onSubmit={handleCreate} />
+          <AddressForm
+            onSubmit={handleCreate}
+            submitButtonText="Save Address"
+          />
 
-          <div className="space-y-4">
+          <div className="space-y-5">
             {addresses.map((address) => (
               <AddressCard
                 key={address.id}
                 address={address}
                 onDelete={handleDelete}
                 onDefault={handleDefault}
+                onEdit={handleEdit}
               />
             ))}
           </div>
